Add -tmx option to generate_map.py to skip image rendering steps.

+++ Scripts/map-maker/generate_map.py
@@ -76,10 +76,12 @@ def main():
     city = CITIES[city_name]
 
     # Make the file structure. Copy from /template
+    tmxOnly = "-tmx" in sys.argv
     output_dir = os.path.join(output_base_dir, city_name)
-    if os.path.exists(output_dir):
-        shutil.rmtree(output_dir)
-    shutil.copytree(template_dir, output_dir)
+    if not tmxOnly:
+        if os.path.exists(output_dir):
+            shutil.rmtree(output_dir)
+        shutil.copytree(template_dir, output_dir)
 
 
     land_image_output_uri = os.path.join(output_dir,"landform.png")
@@ -113,32 +115,42 @@ def main():
     # Render the landform map
     print "RENDERING LANDFORM MAP"
     # render the map to an image
-    land_map_image = mapnik.Image(img_width,img_height)
-    mapnik.render(land_map, land_map_image)
-    land_map_image.save(land_image_output_uri,'png')
+    
+    # If -r option specified attempt to read file rather than regenerate
+    # This allows us to generate only the TMX file
+    land_map_image = None
+    if (tmxOnly and os.path.exists(land_image_output_uri)):
+            print "Loading previously generated landform image"
+            land_map_image = Image.open(land_image_output_uri)
+    
+    if (land_map_image is None):
+        land_map_image = mapnik.Image(img_width,img_height)
+        mapnik.render(land_map, land_map_image)
+        land_map_image.save(land_image_output_uri,'png')
 
-    #Render the refernce map (with streets and junk)
-    print "RENDERING REFERENCE MAP"
-    regular_map = mapnik.Map(img_width,img_height)
-    mapnik.load_map(regular_map, os.path.join(base_path, "stylesheet/osm.xml"))
-    regular_map.srs = merc.params() # ensure the target map projection is mercator
-    regular_map.zoom_to_box(merc_bbox) # fix aspect ratio
-    reference_map_path = os.path.join(output_dir, "reference.png")
-    mapnik.render_to_file(regular_map, reference_map_path)
+    if (not tmxOnly):
+        #Render the refernce map (with streets and junk)
+        print "RENDERING REFERENCE MAP"
+        regular_map = mapnik.Map(img_width,img_height)
+        mapnik.load_map(regular_map, os.path.join(base_path, "stylesheet/osm.xml"))
+        regular_map.srs = merc.params() # ensure the target map projection is mercator
+        regular_map.zoom_to_box(merc_bbox) # fix aspect ratio
+        reference_map_path = os.path.join(output_dir, "reference.png")
+        mapnik.render_to_file(regular_map, reference_map_path)
 
-    # Do any rotating we need with imagemagick
-    if city.get('rotate'):
-        print "PEFORMING ROTATION"
-        # convert reference.png -alpha set \( +clone -background none -rotate 30 \) -gravity center  -compose Src -composite  reference-rotated.png
-        ref_rotated_path = os.path.join(output_dir, "reference-rotated.png")
-        land_rotated_path = os.path.join(output_dir, "landform-rotated.png")
-        reference_rot = Image.open(reference_map_path).rotate(city['rotate'],expand=0)
-        reference_rot.save(ref_rotated_path)
-
-        land_rot = Image.open(land_image_output_uri).rotate(city['rotate'],expand=0)
-        land_rot.save(land_rotated_path)
-
-        land_image_output_uri = land_rotated_path
+        # Do any rotating we need with imagemagick
+        if city.get('rotate'):
+            print "PEFORMING ROTATION"
+            # convert reference.png -alpha set \( +clone -background none -rotate 30 \) -gravity center  -compose Src -composite  reference-rotated.png
+            ref_rotated_path = os.path.join(output_dir, "reference-rotated.png")
+            land_rotated_path = os.path.join(output_dir, "landform-rotated.png")
+            reference_rot = Image.open(reference_map_path).rotate(city['rotate'],expand=0)
+            reference_rot.save(ref_rotated_path)
+    
+            land_rot = Image.open(land_image_output_uri).rotate(city['rotate'],expand=0)
+            land_rot.save(land_rotated_path)
+    
+            land_image_output_uri = land_rotated_path
